Extract REST models to DTO repo
