+++ .idea/.gitignore
@@ -7,3 +7,4 @@ deploymentTargetDropDown.xml
 
 runConfigurations.xml
 SusGameFrontend.iml
+/deploymentTargetSelector.xml
