+++ app/src/main/java/edu/agh/susgame/dto/rest/model/Lobby.kt
@@ -3,13 +3,14 @@
 // IN ORDER TO CHANGE THIS DTO, COMMIT TO 'SusGameDTO' GITHUB REPOSITORY
 // IN ORDER TO UPDATE THIS FILE TO NEWEST VERSION, RUN 'scripts/update-DTO.sh'
 
-package edu.agh.susgame.dto.server
+package edu.agh.susgame.dto.rest.model
 
-import kotlinx.serialization.Serializable
+data class LobbyId(val value: Int)
 
-@Serializable
-data class EdgeDTO(
-    val id: Int,
-    val upgradeCost: Int,
-    val packetsTransported: Int,
+data class Lobby(
+    val id: LobbyId,
+    val name: String,
+    val maxNumOfPlayers: Int,
+    val gameTime: Int,
+    val playersWaiting: Map<PlayerId, Player> = emptyMap(),
 )
