+++ app/src/main/java/edu/agh/susgame/dto/rest/model/Player.kt
@@ -3,11 +3,15 @@
 // IN ORDER TO CHANGE THIS DTO, COMMIT TO 'SusGameDTO' GITHUB REPOSITORY
 // IN ORDER TO UPDATE THIS FILE TO NEWEST VERSION, RUN 'scripts/update-DTO.sh'
 
-package edu.agh.susgame.dto.server
+package edu.agh.susgame.dto.rest.model
 
-import kotlinx.serialization.Serializable
 
-@Serializable
-data class PacketDTO(
-    val id: Int,
+data class PlayerId(val value: Int)
+
+data class PlayerNickname(val value: String)
+
+class Player(
+    val nickname: PlayerNickname,
+    val id: PlayerId,
+    val colorHex: Long,
 )
