+++ app/src/main/java/edu/agh/susgame/dto/socket/ClientSocketMessage.kt
@@ -3,9 +3,9 @@
 // IN ORDER TO CHANGE THIS DTO, COMMIT TO 'SusGameDTO' GITHUB REPOSITORY
 // IN ORDER TO UPDATE THIS FILE TO NEWEST VERSION, RUN 'scripts/update-DTO.sh'
 
-package edu.agh.susgame.dto
+package edu.agh.susgame.dto.socket
 
-import edu.agh.susgame.dto.common.GameStatus
+import edu.agh.susgame.dto.socket.common.GameStatus
 import kotlinx.serialization.Serializable
 
 /**
