+++ app/src/main/java/edu/agh/susgame/dto/socket/common/GameStatus.kt
@@ -3,14 +3,10 @@
 // IN ORDER TO CHANGE THIS DTO, COMMIT TO 'SusGameDTO' GITHUB REPOSITORY
 // IN ORDER TO UPDATE THIS FILE TO NEWEST VERSION, RUN 'scripts/update-DTO.sh'
 
-package edu.agh.susgame.dto.server
+package edu.agh.susgame.dto.socket.common
 
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RouterDTO(
-    val id: Int,
-    val bufferUse: Int,
-    val bufferSize: Int,
-    val upgradeCost: Int,
-)
+enum class GameStatus {
+    WAITING,
+    RUNNING,
+    FINISHED
+}
