+++ app/src/main/java/edu/agh/susgame/dto/socket/server/EdgeDTO.kt
@@ -1,0 +1,15 @@
+// WARNING: THIS FILE WAS CLONED AUTOMATICALLY FROM 'SusGameDTO' GITHUB REPOSITORY
+// IT SHOULD NOT BE EDITED IN ANY WAY
+// IN ORDER TO CHANGE THIS DTO, COMMIT TO 'SusGameDTO' GITHUB REPOSITORY
+// IN ORDER TO UPDATE THIS FILE TO NEWEST VERSION, RUN 'scripts/update-DTO.sh'
+
+package edu.agh.susgame.dto.socket.server
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class EdgeDTO(
+    val id: Int,
+    val upgradeCost: Int,
+    val packetsTransported: Int,
+)
