+++ app/src/main/java/edu/agh/susgame/dto/socket/server/HostDTO.kt
@@ -3,10 +3,12 @@
 // IN ORDER TO CHANGE THIS DTO, COMMIT TO 'SusGameDTO' GITHUB REPOSITORY
 // IN ORDER TO UPDATE THIS FILE TO NEWEST VERSION, RUN 'scripts/update-DTO.sh'
 
-package edu.agh.susgame.dto.common
+package edu.agh.susgame.dto.socket.server
 
-enum class GameStatus {
-    WAITING,
-    RUNNING,
-    FINISHED
-}
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class HostDTO(
+    val id: Int,
+    val packetPath: List<Int>,
+)
