+++ app/src/main/java/edu/agh/susgame/dto/socket/server/RouterDTO.kt
@@ -3,12 +3,14 @@
 // IN ORDER TO CHANGE THIS DTO, COMMIT TO 'SusGameDTO' GITHUB REPOSITORY
 // IN ORDER TO UPDATE THIS FILE TO NEWEST VERSION, RUN 'scripts/update-DTO.sh'
 
-package edu.agh.susgame.dto.server
+package edu.agh.susgame.dto.socket.server
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ServerDTO(
+data class RouterDTO(
     val id: Int,
-    val packetsReceived: Int,
+    val bufferUse: Int,
+    val bufferSize: Int,
+    val upgradeCost: Int,
 )
