+++ app/src/main/java/edu/agh/susgame/dto/socket/server/ServerDTO.kt
@@ -3,12 +3,12 @@
 // IN ORDER TO CHANGE THIS DTO, COMMIT TO 'SusGameDTO' GITHUB REPOSITORY
 // IN ORDER TO UPDATE THIS FILE TO NEWEST VERSION, RUN 'scripts/update-DTO.sh'
 
-package edu.agh.susgame.dto.server
+package edu.agh.susgame.dto.socket.server
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class HostDTO(
+data class ServerDTO(
     val id: Int,
-    val packetPath: List<Int>,
+    val packetsReceived: Int,
 )
