+++ app/src/main/java/edu/agh/susgame/front/model/graph/Edge.kt
@@ -1,7 +1,7 @@
 package edu.agh.susgame.front.model.graph
 
 import androidx.compose.ui.graphics.Color
-import edu.agh.susgame.front.model.PlayerId
+import edu.agh.susgame.dto.rest.model.PlayerId
 
 data class EdgeId(val value: Int)
