+++ app/src/main/java/edu/agh/susgame/front/model/graph/Node.kt
@@ -1,6 +1,6 @@
 package edu.agh.susgame.front.model.graph
 
-import edu.agh.susgame.front.model.PlayerId
+import edu.agh.susgame.dto.rest.model.PlayerId
 import edu.agh.susgame.front.ui.Translation
 import edu.agh.susgame.front.util.Coordinates
