+++ app/src/main/java/edu/agh/susgame/front/service/mock/MockGameService.kt
@@ -1,10 +1,11 @@
 package edu.agh.susgame.front.service.mock
 
-import edu.agh.susgame.front.model.Player
-import edu.agh.susgame.front.model.PlayerNickname
-import edu.agh.susgame.front.model.game.LobbyId
+import edu.agh.susgame.dto.rest.model.LobbyId
+import edu.agh.susgame.dto.rest.model.Player
+import edu.agh.susgame.dto.rest.model.PlayerId
+import edu.agh.susgame.dto.rest.model.PlayerNickname
 import edu.agh.susgame.front.service.interfaces.GameService
-import edu.agh.susgame.front.service.interfaces.GameService.Companion.SimpleMessage
+import edu.agh.susgame.front.service.interfaces.GameService.SimpleMessage
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableSharedFlow
@@ -26,7 +27,7 @@ class MockGameService(private val lobbyService: MockLobbyService) : GameService 
         }
 
     override fun joinLobby(lobbyId: LobbyId, nickname: PlayerNickname): CompletableFuture<Unit> =
-        lobbyService.joinLobby(lobbyId, Player(nickname))
+        lobbyService.joinLobby(lobbyId, Player(nickname, id = PlayerId(0), colorHex = 0x0000FF))
 
     override fun leaveLobby(): CompletableFuture<Unit> =
         when (val lobbyInfo = joinedLobbyInfo) {
