+++ app/src/main/java/edu/agh/susgame/front/service/mock/MockLobbyService.kt
@@ -1,12 +1,13 @@
 package edu.agh.susgame.front.service.mock
 
 import androidx.compose.ui.graphics.Color
+import edu.agh.susgame.dto.rest.model.Lobby
+import edu.agh.susgame.dto.rest.model.LobbyId
+import edu.agh.susgame.dto.rest.model.Player
+import edu.agh.susgame.dto.rest.model.PlayerId
+import edu.agh.susgame.dto.rest.model.PlayerNickname
 import edu.agh.susgame.front.Config
-import edu.agh.susgame.front.model.Player
-import edu.agh.susgame.front.model.PlayerId
-import edu.agh.susgame.front.model.PlayerNickname
-import edu.agh.susgame.front.model.game.Lobby
-import edu.agh.susgame.front.model.game.LobbyId
+import edu.agh.susgame.front.service.interfaces.CreateNewGameResult
 import edu.agh.susgame.front.service.interfaces.LobbyService
 
 import java.util.concurrent.CompletableFuture
@@ -37,7 +38,7 @@ class MockLobbyService(mockDelayMs: Long? = null) : LobbyService {
         gamePin: String,
         maxNumberOfPlayers: Int,
         gameTime: Int,
-    ): CompletableFuture<LobbyService.CreateNewGameResult> =
+    ): CompletableFuture<CreateNewGameResult> =
         CompletableFuture.supplyAsync {
             Thread.sleep(delayMs)
             val lobbyId = LobbyId(freeGameId++)
@@ -50,7 +51,7 @@ class MockLobbyService(mockDelayMs: Long? = null) : LobbyService {
                     gameTime,
                 )
             )
-            LobbyService.CreateNewGameResult.Success(lobbyId)
+            CreateNewGameResult.Success(lobbyId)
         }
 
     /**
@@ -95,11 +96,13 @@ class MockLobbyService(mockDelayMs: Long? = null) : LobbyService {
         // game 1
         val player0 = Player(
             nickname = PlayerNickname("Player_0"),
-            color = Color.Red
+            id = PlayerId(0),
+            colorHex = Color.Red.value.toLong(),
         )
         val player1 = Player(
             nickname = PlayerNickname("Player_1"),
-            color = Color.Green
+            id = PlayerId(1),
+            colorHex = Color.Green.value.toLong(),
         )
 
         val lobbyIdValue1 = LobbyId(freeGameId++)
@@ -118,11 +121,13 @@ class MockLobbyService(mockDelayMs: Long? = null) : LobbyService {
         // game 2
         val player2 = Player(
             nickname = PlayerNickname("Player_0"),
-            color = Color.Red
+            id = PlayerId(0),
+            colorHex = Color.Red.value.toLong(),
         )
         val player3 = Player(
             nickname = PlayerNickname("Player_1"),
-            color = Color.Green
+            id = PlayerId(1),
+            colorHex = Color.Green.value.toLong(),
         )
 
         val lobbyIdValue2 = LobbyId(freeGameId++)
@@ -138,16 +143,9 @@ class MockLobbyService(mockDelayMs: Long? = null) : LobbyService {
         this.joinLobby(lobbyIdValue2, player3)
     }
 
-    private fun getFreePlayerId(): PlayerId {
-        return PlayerId(freePlayerId++)
-    }
-
     private fun lobbyWithPlayerAdded(lobby: Lobby, player: Player): Lobby {
-        val playerId = player.id ?: getFreePlayerId()
-        player.id = playerId
-
         return lobby.copy(
-            playersWaiting = lobby.playersWaiting + (playerId to player),
+            playersWaiting = lobby.playersWaiting + (player.id to player),
         )
     }
 
@@ -161,9 +159,7 @@ class MockLobbyService(mockDelayMs: Long? = null) : LobbyService {
             null -> lobby
             else -> {
                 val playerId = foundPlayer.first
-                val player = foundPlayer.second
 
-                player.resetId()
                 lobby.copy(
                     playersWaiting = lobby.playersWaiting - playerId,
                 )
