+++ app/src/main/java/edu/agh/susgame/front/service/web/WebGameService.kt
@@ -1,8 +1,8 @@
 package edu.agh.susgame.front.service.web
 
-import edu.agh.susgame.dto.ClientSocketMessage
-import edu.agh.susgame.front.model.PlayerNickname
-import edu.agh.susgame.front.model.game.LobbyId
+import edu.agh.susgame.dto.rest.model.LobbyId
+import edu.agh.susgame.dto.rest.model.PlayerNickname
+import edu.agh.susgame.dto.socket.ClientSocketMessage
 import edu.agh.susgame.front.service.interfaces.GameService
 import edu.agh.susgame.front.service.web.rest.AbstractRest
 import edu.agh.susgame.front.service.web.socket.GameWebSocketListener
