--- app/src/main/java/edu/agh/susgame/front/service/web/rest/model/LobbyApi.kt
@@ -1,8 +1,0 @@
-package edu.agh.susgame.front.service.web.rest.model
-
-data class LobbyApi(
-    val id: Int,
-    val name: String,
-    val maxNumberOfPlayers: Int,
-    val players: List<String>,
-)
