+++ app/src/main/java/edu/agh/susgame/front/ui/component/menu/CreateLobbyView.kt
@@ -38,8 +38,8 @@ import androidx.compose.ui.text.input.VisualTransformation
 import androidx.navigation.NavController
 import edu.agh.susgame.front.Config
 import edu.agh.susgame.front.navigation.MenuRoute
+import edu.agh.susgame.front.service.interfaces.CreateNewGameResult
 import edu.agh.susgame.front.service.interfaces.LobbyService
-import edu.agh.susgame.front.service.interfaces.LobbyService.CreateNewGameResult
 import edu.agh.susgame.front.ui.Translation
 import edu.agh.susgame.front.ui.component.common.Header
 import edu.agh.susgame.front.ui.theme.PaddingL
